Moved code in his own file
